add blog page to the project
