+++ src/data.json
@@ -1,0 +1,26 @@
+[
+  {
+    "title": "Quick and Healthy Meals",
+    "src": "https://vegan.com/wp-content/uploads/2018/02/vegan-fitness.jpg",
+    "text": "If you are new to vegan cooking, here are some super easy meals to try. These are all perfect foods for the novice cook, as well as anyone pinched for time!",
+    "id": 1
+  },
+  {
+    "title": "The World’s Greatest Vegan Food Compendium",
+    "src": "https://vegan.com/wp-content/uploads/elementor/thumbs/easy-vegan-foods-qjct056304rmfsiprktmtbc990c06m4cdnwgt4r3c8.jpg",
+    "text": "Get acquainted with every conceivable sort of delicious vegan food. Keep scrolling for thorough coverage of all the best vegan cooking, dining, and grocery possibilities.",
+    "id": 2
+  },
+  {
+    "title": "The World’s Greatest Vegan Food Compendium",
+    "src": "https://vegan.com/wp-content/uploads/2021/06/vegan-snacks.jpg",
+    "text": "Get acquainted with every conceivable sort of delicious vegan food. Keep scrolling for thorough coverage of all the best vegan cooking, dining, and grocery possibilities.",
+    "id": 3
+  },
+  {
+    "title": "The World’s Greatest Vegan Food Compendium",
+    "src": "https://vegan.com/wp-content/uploads/2021/08/vegan-food-2.jpg",
+    "text": "Get acquainted with every conceivable sort of delicious vegan food. Keep scrolling for thorough coverage of all the best vegan cooking, dining, and grocery possibilities.",
+    "id": 4
+  }
+]
